requirements.txt
amqp==5.2.0
async-timeout==4.0.3
bidict==0.23.1
billiard==4.2.1
blinker==1.8.2
celery==5.4.0
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.1
dnspython==2.7.0
eventlet==0.37.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
flask-marshmallow==1.2.1
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
MarkupSafe==3.0.0
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
packaging==24.1
prompt_toolkit==3.0.48
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
redis==5.1.1
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
wsproto==1.2.0
